lnvm: Flush and load bad block table to/from NVM
If bad block table does not exist, it erases and checks the entire
device. From now on, get and set bad block table access the NVM.

The table is saved in the first page of block 1 in lun 0 in each
channel. For now, the maximum bad block table size is 'page_size'.

The OOB area is used to keep information about the table.

The bad block table is composed by an array of bytes, 1 byte per
block and sequential addresses for all channels and luns.

The block is available only if value is 0.
The host may set custom values to a block.